pyproject.toml
[tool.poetry]
name = "telegram bot"
version = "0.1.0"
description = ""
authors = ["McMillan <130283152+McMillankz@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = {extras = ["job-queue"], version = "^20.4"}
aiogram = "^2.25.1"
openai = "^0.27.8"
flask = "^2.3.2"
python-dotenv = "^1.0.0"
authlib = "^1.2.1"
python-chess = "^1.999"
python-lichess = "^0.10"
berserk = "^0.12.8"
stockfish = "^3.28.0"
black = "^23.7.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
datetime = "^5.2"
unicodecsv = "^0.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
